modified query 1 and created helper classes for streaming input from http requests

+++ src/main/java/org/desilvahendricksoftware/debs/Sample.java
@@ -2,11 +2,11 @@ package org.desilvahendricksoftware.debs;
 
 public class Sample {
 
-    int i;
+    long i;
     double voltage;
     double current;
 
-    public Sample(int i, double voltage, double current) {
+    public Sample(long i, double voltage, double current) {
         this.i = i;
         this.voltage = voltage;
         this.current = current;
